+++ src/lib/sort-ships/sort-ships.spec.ts
@@ -1,23 +1,41 @@
-import { sortShips } from '.';
-import { sortById } from './id';
-import { sortByKana } from './kana';
+import { Ship } from '../ship';
+import { sortShips } from './sort';
 import { SortOrder } from './sort-order';
 
-jest.mock('./id');
-jest.mock('./kana');
 describe('sortShips', () => {
-  describe.each`
-    order             | orderStr  | sortFunction
-    ${SortOrder.ID}   | ${'ID'}   | ${sortById}
-    ${SortOrder.KANA} | ${'かな'} | ${sortByKana}
-  `('ソート順が$orderStrのとき', ({ order, sortFunction }) => {
-    beforeEach(() => {
-      (sortFunction as unknown as jest.Mock).mockReturnValue([]);
-      sortShips([], order);
+  describe('ID順を指定した場合', () => {
+    it('ID昇順でソートされる', () => {
+      const ids = [333, 22, 4444, 1];
+      const ships = ids.map((id) => ({ id } as Ship));
+      expect(sortShips(ships, SortOrder.ID)).toEqual([
+        { id: 1 },
+        { id: 22 },
+        { id: 333 },
+        { id: 4444 },
+      ]);
     });
+  });
 
-    it(`${sortFunction.name}が呼ばれる`, () => {
-      expect(sortFunction).toBeCalledWith([]);
+  describe('読み順を指定した場合', () => {
+    it('五十音昇順→アルファベット昇順→その他の文字でソートされる', () => {
+      const kanas = [
+        'かきくけこ',
+        'さしすせそ',
+        'Верный',
+        'fghij',
+        'abcde',
+        'あいうえお',
+      ];
+      const ships = kanas.map((kana) => ({ kana } as Ship));
+
+      expect(sortShips(ships, SortOrder.KANA)).toEqual([
+        { kana: 'あいうえお' },
+        { kana: 'かきくけこ' },
+        { kana: 'さしすせそ' },
+        { kana: 'abcde' },
+        { kana: 'fghij' },
+        { kana: 'Верный' },
+      ]);
     });
   });
